Auth basic: Cache credentials if auth_basic_user_file is a static text.
Doing file I/O every HTTP request slows down performance.  If the specified
credential is a static file path, reading it at configuration time and
all worker processes can reuse it.  If the path contains variables, we keep
the same behavior because it needs to be calculated in every request.